refactor(administration): merge emit-native-setup into emit-composition-api-script
The two-file split existed while the second half was swappable between a
native setup and an Options API backoff emitter. With only the native setup
emitter left, the split just spreads one output order over two files.
